+++ src/__tests__/pages/develop-your-curriculum.test.tsx
@@ -4,6 +4,7 @@ import { BlogPostPreview, CurriculumPage } from "../../node-lib/cms";
 import renderWithProviders from "../__helpers__/renderWithProviders";
 import { mockSeo, portableTextFromString } from "../__helpers__/cms";
 import Curriculum from "../../pages/develop-your-curriculum";
+import renderWithSeo from "../__helpers__/renderWithSeo";
 
 const testCurriculumPageData: CurriculumPage = {
   id: "01",
@@ -61,7 +62,7 @@ const testCurriculumPageData: CurriculumPage = {
 
 const getPageData = jest.fn(() => testCurriculumPageData);
 
-describe("pages/curriculum.tsx", () => {
+describe("pages/develop-your-curriculum.tsx", () => {
   beforeEach(() => {
     jest.clearAllMocks();
     jest.resetModules();
@@ -74,9 +75,7 @@ describe("pages/curriculum.tsx", () => {
   });
 
   it("Renders correct title ", async () => {
-    renderWithProviders(
-      <Curriculum pageData={testCurriculumPageData} isPreviewMode={false} />
-    );
+    renderWithProviders(<Curriculum pageData={testCurriculumPageData} />);
 
     await waitFor(() => {
       expect(screen.getByRole("heading", { level: 1 }).textContent).toBe(
@@ -86,9 +85,7 @@ describe("pages/curriculum.tsx", () => {
   });
 
   it("renders the blog posts", async () => {
-    renderWithProviders(
-      <Curriculum pageData={testCurriculumPageData} isPreviewMode={false} />
-    );
+    renderWithProviders(<Curriculum pageData={testCurriculumPageData} />);
     const { posts } = testCurriculumPageData.elements;
 
     await waitFor(() => {
@@ -99,6 +96,16 @@ describe("pages/curriculum.tsx", () => {
 
       expect(links[0]).toHaveAttribute("href", "/blog/some-post");
       expect(links[1]).toHaveAttribute("href", "/blog/some-other-post");
+    });
+  });
+
+  describe.skip("SEO", () => {
+    it("renders the correct SEO details", async () => {
+      const { seo } = renderWithSeo(
+        <Curriculum pageData={testCurriculumPageData} />
+      );
+
+      expect(seo).toEqual({});
     });
   });