+++ src/node-lib/cms/sanity-client/schemas/base.ts
@@ -25,6 +25,7 @@ export const imageAssetSchema = z.object({
 
 export const imageSchema = z.object({
   altText: z.string().nonempty().nullish(),
+  isPresentational: z.boolean().nullish(),
   asset: imageAssetSchema,
 });