+++ src/pages/lp/[landingPageSlug].tsx
@@ -1,6 +1,6 @@
 import { GetStaticPaths, GetStaticProps, NextPage } from "next";
 
-import { DEFAULT_SEO_PROPS } from "../../browser-lib/seo/Seo";
+import { getSeoProps } from "../../browser-lib/seo/getSeoProps";
 import Grid, { GridArea } from "../../components/Grid";
 import Layout from "../../components/Layout";
 import MaxWidth from "../../components/MaxWidth/MaxWidth";
@@ -8,12 +8,11 @@ import CMSClient, { LandingPage } from "../../node-lib/cms";
 
 export type LandingPageProps = {
   pageData: LandingPage;
-  isPreviewMode: boolean;
 };
 
-const Landing: NextPage<LandingPageProps> = ({ pageData, isPreviewMode }) => {
+const Landing: NextPage<LandingPageProps> = ({ pageData }) => {
   return (
-    <Layout seoProps={DEFAULT_SEO_PROPS} isPreviewMode={isPreviewMode}>
+    <Layout seoProps={getSeoProps(pageData.seo)}>
       <MaxWidth>
         <Grid>
           <GridArea $colSpan={[12, 12, 12]}>
@@ -62,7 +61,6 @@ export const getStaticProps: GetStaticProps<
   return {
     props: {
       pageData: landingPageResult,
-      isPreviewMode,
     },
     revalidate: 10,
   };